+++ src/main/java/com/coolerpromc/easybrewing/compat/cobblemon/CobblemonBottleIngredientCheck.java
@@ -31,4 +31,22 @@ public class CobblemonBottleIngredientCheck {
         }
         return false;
     }
+
+    public static boolean isCobblemonIngredient(ItemStack stack, Level level){
+        if(ModList.get().isLoaded("cobblemon")){
+            try{
+                RecipeType<BrewingStandRecipe> brewingStandType = (RecipeType<BrewingStandRecipe>) BuiltInRegistries.RECIPE_TYPE.get(ResourceLocation.fromNamespaceAndPath("cobblemon", "brewing_stand"));
+                if (brewingStandType != null){
+                    List<Ingredient> ingredients = level.getRecipeManager().getAllRecipesFor(brewingStandType).stream().map(RecipeHolder::value).map(BrewingStandRecipe::getInput).toList();
+                    for (Ingredient ingredient : ingredients){
+                        if (ingredient.test(stack)) return true;
+                    }
+                }
+            }
+            catch (Exception e){
+                return false;
+            }
+        }
+        return false;
+    }
 }